src/backend/requirements.txt
anyio==3.5.0
asgiref==3.5.1
autopep8==1.6.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
fastapi==0.75.2
h11==0.13.0
idna==3.3
install==1.3.5
Jinja2==3.1.2
MarkupSafe==2.1.1
pendulum==2.1.2
pycodestyle==2.8.0
pydantic==1.9.0
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytzdata==2020.1
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
